[Build] Universally support test muting
In scope of: KT-88863
^KT-88863 fixed
