+++ repo/gradle-build-conventions/project-tests-convention/src/main/kotlin/project-tests-convention.gradle.kts
@@ -12,3 +12,4 @@ configureTestRetries()
 configureTestCacheDisabling()
 configureDefaultJvmArguments()
 configureTestInputs()
+configureTestMuting()
